refactor: Add folders (#15)
