+++ lib/auth/model.dart
