+++ lib/main_page.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
-import 'package:flutter_neighclova/edit_info.dart';
-import 'package:flutter_neighclova/mypage.dart';
-import 'package:flutter_neighclova/news.dart';
-import 'package:flutter_neighclova/register_info.dart';
+import 'package:flutter_neighclova/place/edit_info.dart';
+import 'package:flutter_neighclova/mypage/mypage.dart';
+import 'package:flutter_neighclova/news/news.dart';
+import 'package:flutter_neighclova/place/register_info.dart';
 import 'dart:ui';
 
 import 'package:flutter_neighclova/tabview.dart';
@@ -17,35 +17,35 @@ class MainPage extends StatefulWidget {
 }
 
 class _MainPageState extends State<MainPage> {
-
 //가게들 넘겨받기
- List<Store> stores = [];
- int selectedIndex = 0;
+  List<Store> stores = [];
+  int selectedIndex = 0;
 
- final GlobalKey _containerKey = GlobalKey();
- double _containerHeight = 0;
+  final GlobalKey _containerKey = GlobalKey();
+  double _containerHeight = 0;
 
- final GlobalKey _wrapKey = GlobalKey();
- double _wrapHeight = 0;
+  final GlobalKey _wrapKey = GlobalKey();
+  double _wrapHeight = 0;
 
- final GlobalKey _goodFeedbackKey = GlobalKey();
- double _goodFeedbackHeight = 0;
+  final GlobalKey _goodFeedbackKey = GlobalKey();
+  double _goodFeedbackHeight = 0;
 
- final GlobalKey _badFeedbackKey = GlobalKey();
- double _badFeedbackHeight = 0;
+  final GlobalKey _badFeedbackKey = GlobalKey();
+  double _badFeedbackHeight = 0;
 
- //키워드
- List<String> keywords = [];
+  //키워드
+  List<String> keywords = [];
 
- //피드백
- String goodFeedback = '음식이 맛있고 사장님이 친절해요.';
- String badFeedback = '음식에 먼지가 나왔어요. 위생에 유의해 주세요.';
+  //피드백
+  String goodFeedback = '음식이 맛있고 사장님이 친절해요.';
+  String badFeedback = '음식에 먼지가 나왔어요. 위생에 유의해 주세요.';
 
- //소식 생성 날짜
- int afterGenerateNews = 0;
+  //소식 생성 날짜
+  int afterGenerateNews = 0;
 
- void _updateContainerHeight(StateSetter bottomState) {
-    final RenderBox? renderBox = _containerKey.currentContext?.findRenderObject() as RenderBox?;
+  void _updateContainerHeight(StateSetter bottomState) {
+    final RenderBox? renderBox =
+        _containerKey.currentContext?.findRenderObject() as RenderBox?;
     if (renderBox != null) {
       setState(() {
         _containerHeight = renderBox.size.height;
@@ -55,7 +55,8 @@ class _MainPageState extends State<MainPage> {
   }
 
   void _updateWrapHeight() {
-    final RenderBox? renderBox = _wrapKey.currentContext?.findRenderObject() as RenderBox?;
+    final RenderBox? renderBox =
+        _wrapKey.currentContext?.findRenderObject() as RenderBox?;
     if (renderBox != null) {
       setState(() {
         _wrapHeight = renderBox.size.height;
@@ -65,7 +66,8 @@ class _MainPageState extends State<MainPage> {
   }
 
   void _updateGoodFeedbackHeight() {
-    final RenderBox? renderBox = _goodFeedbackKey.currentContext?.findRenderObject() as RenderBox?;
+    final RenderBox? renderBox =
+        _goodFeedbackKey.currentContext?.findRenderObject() as RenderBox?;
     if (renderBox != null) {
       setState(() {
         _goodFeedbackHeight = renderBox.size.height;
@@ -75,7 +77,8 @@ class _MainPageState extends State<MainPage> {
   }
 
   void _updateBadFeedbackHeight() {
-    final RenderBox? renderBox = _badFeedbackKey.currentContext?.findRenderObject() as RenderBox?;
+    final RenderBox? renderBox =
+        _badFeedbackKey.currentContext?.findRenderObject() as RenderBox?;
     if (renderBox != null) {
       setState(() {
         _badFeedbackHeight = renderBox.size.height;
@@ -89,14 +92,11 @@ class _MainPageState extends State<MainPage> {
     super.initState();
 
     WidgetsBinding.instance.addPostFrameCallback((_) {
-      if (!isRegistered)
-      {
+      if (!isRegistered) {
         Navigator.push(
-          context,
-          MaterialPageRoute(
-            builder: (BuildContext context) =>
-            RegisterInfo())
-        );
+            context,
+            MaterialPageRoute(
+                builder: (BuildContext context) => RegisterInfo()));
       }
       _updateWrapHeight();
       _updateGoodFeedbackHeight();
@@ -106,7 +106,11 @@ class _MainPageState extends State<MainPage> {
     storeName = '소곤 식당';
 
     //전체 가게 이름
-    stores.addAll([Store('소곤 식당', 'assets/storeImg.png'), Store('소곤 카페', ''), Store('망한 식당', '')]);
+    stores.addAll([
+      Store('소곤 식당', 'assets/storeImg.png'),
+      Store('소곤 카페', ''),
+      Store('망한 식당', '')
+    ]);
 
     //바텀시트 가게 선택 상태
 
@@ -120,6 +124,7 @@ class _MainPageState extends State<MainPage> {
     //소식 생성 날짜 세팅
     afterGenerateNews = 10;
   }
+
   bool isRegistered = false;
   String storeName = '';
 
@@ -153,23 +158,23 @@ class _MainPageState extends State<MainPage> {
                 width: double.infinity,
                 height: 67,
                 decoration: BoxDecoration(
-                  borderRadius: BorderRadius.circular(14),
-                  color: Colors.white,
-                  boxShadow: [
-                    BoxShadow(
-                      color: Colors.grey.withOpacity(0.5),
-                      blurRadius: 24,
-                      offset: Offset(0, 8),
-                    )
-                  ]
-                ),
+                    borderRadius: BorderRadius.circular(14),
+                    color: Colors.white,
+                    boxShadow: [
+                      BoxShadow(
+                        color: Colors.grey.withOpacity(0.5),
+                        blurRadius: 24,
+                        offset: Offset(0, 8),
+                      )
+                    ]),
                 child: Padding(
                   padding: EdgeInsets.fromLTRB(16, 20, 16, 20),
                   child: Row(
                     mainAxisAlignment: MainAxisAlignment.spaceBetween,
                     crossAxisAlignment: CrossAxisAlignment.center,
                     children: [
-                      Text(storeName,
+                      Text(
+                        storeName,
                         style: TextStyle(
                           fontSize: 20,
                           fontWeight: FontWeight.bold,
@@ -179,176 +184,221 @@ class _MainPageState extends State<MainPage> {
                       TextButton(
                         onPressed: () {
                           showModalBottomSheet(
-                            context: context,
-                            isScrollControlled: true,
-                            builder: (BuildContext context) {
-                              return StatefulBuilder(builder: (BuildContext context, StateSetter bottomState) {
-                                WidgetsBinding.instance.addPostFrameCallback((_) => _updateContainerHeight(bottomState));
-                                void onButtonPressed(int index) {
-                                  storeName = stores[index].storeName;
-                                  bottomState(() {
-                                    setState(() {
-                                      selectedIndex = index;
+                              context: context,
+                              isScrollControlled: true,
+                              builder: (BuildContext context) {
+                                return StatefulBuilder(builder:
+                                    (BuildContext context,
+                                        StateSetter bottomState) {
+                                  WidgetsBinding.instance.addPostFrameCallback(
+                                      (_) =>
+                                          _updateContainerHeight(bottomState));
+                                  void onButtonPressed(int index) {
+                                    storeName = stores[index].storeName;
+                                    bottomState(() {
+                                      setState(() {
+                                        selectedIndex = index;
+                                      });
                                     });
-                                  });
-                                  Navigator.pop(context);
-                                  (context as Element).reassemble();
-                                }
-                                return Container(
-                                  height: 114 + 64.0 * (stores.length + 1),
-                                  decoration: BoxDecoration(
-                                    gradient: LinearGradient(
-                                      begin: Alignment.topLeft,
-                                      end: Alignment.bottomRight,
-                                      colors: [
-                                        Color(0xffF9FFE1),
-                                        Color(0xffFFEDED),
-                                        Color(0xffFDF6FF),
-                                        Color(0xffFCFFE9),
-                                        Color(0xffEAFFF5),
-                                        Color(0xffF0FBFF),
-                                      ]
+                                    Navigator.pop(context);
+                                    (context as Element).reassemble();
+                                  }
+
+                                  return Container(
+                                    height: 114 + 64.0 * (stores.length + 1),
+                                    decoration: BoxDecoration(
+                                      gradient: LinearGradient(
+                                          begin: Alignment.topLeft,
+                                          end: Alignment.bottomRight,
+                                          colors: [
+                                            Color(0xffF9FFE1),
+                                            Color(0xffFFEDED),
+                                            Color(0xffFDF6FF),
+                                            Color(0xffFCFFE9),
+                                            Color(0xffEAFFF5),
+                                            Color(0xffF0FBFF),
+                                          ]),
+                                      borderRadius: BorderRadius.circular(13),
                                     ),
-                                    borderRadius: BorderRadius.circular(13),
-                                  ),
-                                  child: Padding(
-                                    padding: EdgeInsets.fromLTRB(20, 10, 20, 23),
-                                    child: Column(
-                                      mainAxisSize: MainAxisSize.min,
-                                      children: [
-                                        Container(
-                                          height: 5,
-                                          width: 50,
-                                          decoration: BoxDecoration(
-                                            color: Colors.grey[400],
-                                            borderRadius: BorderRadius.circular(10),
+                                    child: Padding(
+                                      padding:
+                                          EdgeInsets.fromLTRB(20, 10, 20, 23),
+                                      child: Column(
+                                        mainAxisSize: MainAxisSize.min,
+                                        children: [
+                                          Container(
+                                            height: 5,
+                                            width: 50,
+                                            decoration: BoxDecoration(
+                                              color: Colors.grey[400],
+                                              borderRadius:
+                                                  BorderRadius.circular(10),
+                                            ),
                                           ),
-                                        ),
-                                        Padding(padding: EdgeInsets.only(top: 17),),
-                                        Container(
-                                          key: _containerKey,
-                                          decoration: BoxDecoration(
-                                            color: Colors.white,
-                                            borderRadius: BorderRadius.circular(12),
+                                          Padding(
+                                            padding: EdgeInsets.only(top: 17),
                                           ),
-                                          child: Flexible(
-                                            child: ListView.builder(
-                                              shrinkWrap: true,
-                                              itemCount: stores.length + 1,
-                                              itemBuilder: (context, index) {
-                                                if (index == stores.length) {
-                                                  return ListTile(
-                                                    leading: CircleAvatar(
-                                                      backgroundColor: Color(0xffF2F2F2),
-                                                      child: ClipOval(
-                                                        child: Icon(Icons.add, color: Color(0xff656565)),
+                                          Container(
+                                            key: _containerKey,
+                                            decoration: BoxDecoration(
+                                              color: Colors.white,
+                                              borderRadius:
+                                                  BorderRadius.circular(12),
+                                            ),
+                                            child: Flexible(
+                                              child: ListView.builder(
+                                                shrinkWrap: true,
+                                                itemCount: stores.length + 1,
+                                                itemBuilder: (context, index) {
+                                                  if (index == stores.length) {
+                                                    return ListTile(
+                                                      leading: CircleAvatar(
+                                                        backgroundColor:
+                                                            Color(0xffF2F2F2),
+                                                        child: ClipOval(
+                                                          child: Icon(Icons.add,
+                                                              color: Color(
+                                                                  0xff656565)),
+                                                        ),
                                                       ),
-                                                    ),
-                                                    title: TextButton(
-                                                      onPressed: () {
-                                                        Navigator.pushReplacement(
-                                                          context,
-                                                          MaterialPageRoute(
-                                                            builder: (context) => RegisterInfo()
-                                                          ),
-                                                        );
-                                                      },
-                                                      child: Align(
-                                                        alignment: Alignment.centerLeft,
-                                                        child: Text(
-                                                          '내 업체 추가',
-                                                          textAlign: TextAlign.left,
-                                                          style: TextStyle(
-                                                            fontSize: 16,
+                                                      title: TextButton(
+                                                        onPressed: () {
+                                                          Navigator
+                                                              .pushReplacement(
+                                                            context,
+                                                            MaterialPageRoute(
+                                                                builder:
+                                                                    (context) =>
+                                                                        RegisterInfo()),
+                                                          );
+                                                        },
+                                                        child: Align(
+                                                          alignment: Alignment
+                                                              .centerLeft,
+                                                          child: Text(
+                                                            '내 업체 추가',
+                                                            textAlign:
+                                                                TextAlign.left,
+                                                            style: TextStyle(
+                                                              fontSize: 16,
+                                                            ),
                                                           ),
                                                         ),
                                                       ),
-                                                    ),
-                                                    trailing: null,
-                                                  );
-                                                }
-                                                else {
-                                                  return ListTile(
-                                                    leading: CircleAvatar(
-                                                      backgroundColor: Colors.grey[400],
-                                                      child: ClipOval(
-                                                        child: stores[index].imgUrl != ''
-                                                        ? Image.asset(stores[index].imgUrl,
-                                                            fit: BoxFit.cover,
-                                                            width: double.infinity,
-                                                            height: double.infinity,
-                                                          )
-                                                        : Icon(
-                                                            Icons.person,
-                                                            color: Colors.white,
-                                                          )
+                                                      trailing: null,
+                                                    );
+                                                  } else {
+                                                    return ListTile(
+                                                      leading: CircleAvatar(
+                                                        backgroundColor:
+                                                            Colors.grey[400],
+                                                        child: ClipOval(
+                                                            child: stores[index]
+                                                                        .imgUrl !=
+                                                                    ''
+                                                                ? Image.asset(
+                                                                    stores[index]
+                                                                        .imgUrl,
+                                                                    fit: BoxFit
+                                                                        .cover,
+                                                                    width: double
+                                                                        .infinity,
+                                                                    height: double
+                                                                        .infinity,
+                                                                  )
+                                                                : Icon(
+                                                                    Icons
+                                                                        .person,
+                                                                    color: Colors
+                                                                        .white,
+                                                                  )),
                                                       ),
-                                                    ),
-                                                    title: TextButton(
-                                                      onPressed: () => onButtonPressed(index),
-                                                      child: Align(
-                                                        alignment: Alignment.centerLeft,
-                                                        child: Text(stores[index].storeName,
-                                                        textAlign: TextAlign.left,
-                                                        style: TextStyle(
-                                                          fontSize: 16,
-                                                        ),),
-                                                      )
-                                                    ),
-                                                    trailing: selectedIndex == index ? Icon(Icons.check_circle, color: Color(0xff468F4F)) : null,
-                                                  );
-                                                }
+                                                      title: TextButton(
+                                                          onPressed: () =>
+                                                              onButtonPressed(
+                                                                  index),
+                                                          child: Align(
+                                                            alignment: Alignment
+                                                                .centerLeft,
+                                                            child: Text(
+                                                              stores[index]
+                                                                  .storeName,
+                                                              textAlign:
+                                                                  TextAlign
+                                                                      .left,
+                                                              style: TextStyle(
+                                                                fontSize: 16,
+                                                              ),
+                                                            ),
+                                                          )),
+                                                      trailing: selectedIndex ==
+                                                              index
+                                                          ? Icon(
+                                                              Icons
+                                                                  .check_circle,
+                                                              color: Color(
+                                                                  0xff468F4F))
+                                                          : null,
+                                                    );
+                                                  }
+                                                },
+                                              ),
+                                            ),
+                                          ),
+                                          Padding(
+                                            padding: EdgeInsets.only(top: 11),
+                                          ),
+                                          SizedBox(
+                                            width: double.infinity,
+                                            child: ElevatedButton(
+                                              onPressed: () {
+                                                Navigator.push(
+                                                    context,
+                                                    MaterialPageRoute(
+                                                        builder: (BuildContext
+                                                                context) =>
+                                                            MyPage()));
                                               },
-                                            ),
-                                          ),
-                                        ),
-                                        Padding(padding: EdgeInsets.only(top: 11),),
-                                        SizedBox(
-                                          width: double.infinity,
-                                          child: ElevatedButton(
-                                            onPressed: () {
-                                              Navigator.push(
-                                                context,
-                                                MaterialPageRoute(
-                                                  builder: (BuildContext context) =>
-                                                    MyPage())
-                                              );
-                                            },
-                                            style: ElevatedButton.styleFrom(
-                                              backgroundColor: Colors.transparent,
-                                              shape: RoundedRectangleBorder(
-                                                borderRadius: BorderRadius.circular(20),
+                                              style: ElevatedButton.styleFrom(
+                                                backgroundColor:
+                                                    Colors.transparent,
+                                                shape: RoundedRectangleBorder(
+                                                  borderRadius:
+                                                      BorderRadius.circular(20),
+                                                ),
+                                                side: BorderSide(
+                                                  color: Color(0xff878787),
+                                                  width: 0.3,
+                                                ),
+                                                elevation: 0,
                                               ),
-                                              side: BorderSide(
-                                                color: Color(0xff878787),
-                                                width: 0.3,
+                                              child: Text(
+                                                '내 업체 정보 페이지로 이동',
+                                                style: TextStyle(
+                                                  fontSize: 14,
+                                                ),
                                               ),
-                                              elevation: 0,
                                             ),
-                                            child: Text('내 업체 정보 페이지로 이동',
-                                            style: TextStyle(
-                                              fontSize: 14,
-                                            ),),
-                                          ),
-                                        )
-                                      ],
+                                          )
+                                        ],
+                                      ),
                                     ),
-                                  ),
-                                );
+                                  );
+                                });
                               });
-                            }
-                          );
                         },
                         child: Row(
                           mainAxisSize: MainAxisSize.min,
                           children: [
-                            Text('업체 변경하기',
+                            Text(
+                              '업체 변경하기',
                               style: TextStyle(
                                 fontSize: 14,
                                 color: Color(0xff949494),
                               ),
                             ),
-                            Icon(Icons.arrow_forward_ios,
+                            Icon(
+                              Icons.arrow_forward_ios,
                               color: Color(0xff949494),
                               size: 20,
                             ),
@@ -357,9 +407,12 @@ class _MainPageState extends State<MainPage> {
                         style: ButtonStyle(
                           alignment: Alignment.centerRight,
                           padding: MaterialStateProperty.all(EdgeInsets.all(0)),
-                          backgroundColor: MaterialStateProperty.all(Colors.white),
-                          shadowColor: MaterialStateProperty.all(Colors.transparent),
-                          overlayColor: MaterialStateProperty.all(Colors.transparent),
+                          backgroundColor:
+                              MaterialStateProperty.all(Colors.white),
+                          shadowColor:
+                              MaterialStateProperty.all(Colors.transparent),
+                          overlayColor:
+                              MaterialStateProperty.all(Colors.transparent),
                         ),
                       ),
                     ],
@@ -371,16 +424,15 @@ class _MainPageState extends State<MainPage> {
                 width: double.infinity,
                 height: 129 + _wrapHeight,
                 decoration: BoxDecoration(
-                  borderRadius: BorderRadius.circular(14),
-                  color: Colors.white,
-                  boxShadow: [
-                    BoxShadow(
-                      color: Colors.grey.withOpacity(0.5),
-                      blurRadius: 24,
-                      offset: Offset(0, 8),
-                    )
-                  ]
-                ),
+                    borderRadius: BorderRadius.circular(14),
+                    color: Colors.white,
+                    boxShadow: [
+                      BoxShadow(
+                        color: Colors.grey.withOpacity(0.5),
+                        blurRadius: 24,
+                        offset: Offset(0, 8),
+                      )
+                    ]),
                 child: Padding(
                   padding: EdgeInsets.fromLTRB(16, 20, 16, 6),
                   child: Column(
@@ -388,7 +440,8 @@ class _MainPageState extends State<MainPage> {
                     children: [
                       Align(
                         alignment: Alignment.topLeft,
-                        child: Text('우리 가게 대표 키워드',
+                        child: Text(
+                          '우리 가게 대표 키워드',
                           style: TextStyle(
                             fontSize: 20,
                             fontWeight: FontWeight.bold,
@@ -399,7 +452,8 @@ class _MainPageState extends State<MainPage> {
                       Padding(padding: EdgeInsets.only(top: 4)),
                       Align(
                         alignment: Alignment.topLeft,
-                        child: Text('CLOVA AI가 매장 리뷰를 분석해서 뽑은 키워드예요.',
+                        child: Text(
+                          'CLOVA AI가 매장 리뷰를 분석해서 뽑은 키워드예요.',
                           style: TextStyle(
                             fontSize: 12,
                             color: Color(0xff949494),
@@ -426,16 +480,15 @@ class _MainPageState extends State<MainPage> {
                 width: double.infinity,
                 height: 145 + _goodFeedbackHeight + _badFeedbackHeight,
                 decoration: BoxDecoration(
-                  borderRadius: BorderRadius.circular(14),
-                  color: Colors.white,
-                  boxShadow: [
-                    BoxShadow(
-                      color: Colors.grey.withOpacity(0.5),
-                      blurRadius: 24,
-                      offset: Offset(0, 8),
-                    )
-                  ]
-                ),
+                    borderRadius: BorderRadius.circular(14),
+                    color: Colors.white,
+                    boxShadow: [
+                      BoxShadow(
+                        color: Colors.grey.withOpacity(0.5),
+                        blurRadius: 24,
+                        offset: Offset(0, 8),
+                      )
+                    ]),
                 child: Padding(
                   padding: EdgeInsets.fromLTRB(16, 20, 16, 6),
                   child: Column(
@@ -443,7 +496,8 @@ class _MainPageState extends State<MainPage> {
                     children: [
                       Align(
                         alignment: Alignment.topLeft,
-                        child: Text('피드백',
+                        child: Text(
+                          '피드백',
                           style: TextStyle(
                             fontSize: 20,
                             fontWeight: FontWeight.bold,
@@ -454,7 +508,8 @@ class _MainPageState extends State<MainPage> {
                       Padding(padding: EdgeInsets.only(top: 4)),
                       Align(
                         alignment: Alignment.topLeft,
-                        child: Text('CLOVA AI가 매장 리뷰를 분석해서 제공하는 피드백이에요.',
+                        child: Text(
+                          'CLOVA AI가 매장 리뷰를 분석해서 제공하는 피드백이에요.',
                           style: TextStyle(
                             fontSize: 12,
                             color: Color(0xff949494),
@@ -464,7 +519,8 @@ class _MainPageState extends State<MainPage> {
                       Padding(padding: EdgeInsets.only(top: 22)),
                       Align(
                         alignment: Alignment.topLeft,
-                        child: Text(goodFeedback,
+                        child: Text(
+                          goodFeedback,
                           key: _goodFeedbackKey,
                           style: TextStyle(
                             fontSize: 14,
@@ -474,7 +530,8 @@ class _MainPageState extends State<MainPage> {
                       Padding(padding: EdgeInsets.only(top: 16)),
                       Align(
                         alignment: Alignment.topLeft,
-                        child: Text(badFeedback,
+                        child: Text(
+                          badFeedback,
                           key: _badFeedbackKey,
                           style: TextStyle(
                             fontSize: 14,
@@ -491,16 +548,15 @@ class _MainPageState extends State<MainPage> {
                 width: double.infinity,
                 height: 188,
                 decoration: BoxDecoration(
-                  borderRadius: BorderRadius.circular(14),
-                  color: Colors.white,
-                  boxShadow: [
-                    BoxShadow(
-                      color: Colors.grey.withOpacity(0.5),
-                      blurRadius: 24,
-                      offset: Offset(0, 8),
-                    )
-                  ]
-                ),
+                    borderRadius: BorderRadius.circular(14),
+                    color: Colors.white,
+                    boxShadow: [
+                      BoxShadow(
+                        color: Colors.grey.withOpacity(0.5),
+                        blurRadius: 24,
+                        offset: Offset(0, 8),
+                      )
+                    ]),
                 child: Padding(
                   padding: EdgeInsets.fromLTRB(16, 20, 16, 14),
                   child: Column(
@@ -508,7 +564,8 @@ class _MainPageState extends State<MainPage> {
                     children: [
                       Align(
                         alignment: Alignment.topLeft,
-                        child: Text('소식',
+                        child: Text(
+                          '소식',
                           style: TextStyle(
                             fontSize: 20,
                             fontWeight: FontWeight.bold,
@@ -520,35 +577,34 @@ class _MainPageState extends State<MainPage> {
                       Align(
                         alignment: Alignment.topLeft,
                         child: RichText(
-                          text: TextSpan(
-                            text: '새로운 소식 글을 작성한지 ',
-                            style: TextStyle(
-                              fontSize: 14,
-                              color: Color(0xff404040),
-                            ), 
-                            children: <TextSpan>[
+                            text: TextSpan(
+                                text: '새로운 소식 글을 작성한지 ',
+                                style: TextStyle(
+                                  fontSize: 14,
+                                  color: Color(0xff404040),
+                                ),
+                                children: <TextSpan>[
                               TextSpan(
                                 text: afterGenerateNews.toString() + '일',
                                 style: TextStyle(
                                   fontSize: 14,
                                   color: Color(0xff03AA5A),
-                                ),  
+                                ),
                               ),
                               TextSpan(
                                 text: '이 지났어요.',
                                 style: TextStyle(
                                   fontSize: 14,
                                   color: Color(0xff404040),
-                                ), 
+                                ),
                               ),
-                            ]
-                          )
-                        ),
+                            ])),
                       ),
                       Padding(padding: EdgeInsets.only(top: 8)),
                       Align(
                         alignment: Alignment.topLeft,
-                        child: Text('매장 홍보를 위해 소식 글을 업로드해 보세요.',
+                        child: Text(
+                          '매장 홍보를 위해 소식 글을 업로드해 보세요.',
                           style: TextStyle(
                             fontSize: 14,
                             color: Color(0xff404040),
@@ -561,11 +617,11 @@ class _MainPageState extends State<MainPage> {
                         child: ElevatedButton(
                           onPressed: () {
                             final tabViewState = TabView.of(context);
-                              if (tabViewState != null) {
-                                tabViewState.navigateToNewsPage();
-                              } else {
-                                print('탭바 오류');
-                              }
+                            if (tabViewState != null) {
+                              tabViewState.navigateToNewsPage();
+                            } else {
+                              print('탭바 오류');
+                            }
                           },
                           style: ElevatedButton.styleFrom(
                             backgroundColor: Colors.transparent,
@@ -578,11 +634,11 @@ class _MainPageState extends State<MainPage> {
                             ),
                             elevation: 0,
                           ),
-                          child: Text('소식 글 생성하기',
-                          style: TextStyle(
-                            fontSize: 14,
-                            color: Color(0xff03AA5A)
-                          ),),
+                          child: Text(
+                            '소식 글 생성하기',
+                            style: TextStyle(
+                                fontSize: 14, color: Color(0xff03AA5A)),
+                          ),
                         ),
                       ),
                     ],
@@ -603,13 +659,11 @@ class _MainPageState extends State<MainPage> {
         width: keywords[index].length * 15.0,
         child: Align(
           alignment: Alignment.center,
-          child: Text(
-            keywords[index],
-            style: TextStyle(
-            color: Colors.white,
-            fontSize: 14,
-            )
-          ),
+          child: Text(keywords[index],
+              style: TextStyle(
+                color: Colors.white,
+                fontSize: 14,
+              )),
         ),
       ),
       backgroundColor: Color(0xff03AA5A),
@@ -622,7 +676,7 @@ class _MainPageState extends State<MainPage> {
   }
 }
 
-class Store{
+class Store {
   String storeName;
   String imgUrl;
   Store(this.storeName, this.imgUrl);
