+++ lib/news/news.dart
@@ -2,8 +2,8 @@ import 'dart:convert';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
-import 'package:flutter_neighclova/generate_news.dart';
-import 'package:flutter_neighclova/news_card.dart';
+import 'package:flutter_neighclova/news/generate_news.dart';
+import 'package:flutter_neighclova/news/news_card.dart';
 
 class NewsPage extends StatefulWidget {
   const NewsPage({Key? key}) : super(key: key);
@@ -27,12 +27,11 @@ class _NewsPageState extends State<NewsPage> {
           ),
         ),
         title: Text('소식 생성',
-          style: TextStyle(
-            fontWeight: FontWeight.bold,
-            color: Color(0xff404040),
-            fontSize: 20,
-          )
-        ),
+            style: TextStyle(
+              fontWeight: FontWeight.bold,
+              color: Color(0xff404040),
+              fontSize: 20,
+            )),
         centerTitle: true,
         automaticallyImplyLeading: false,
       ),
@@ -45,17 +44,17 @@ class _NewsPageState extends State<NewsPage> {
           );
         },
         separatorBuilder: (BuildContext context, int index) {
-          return SizedBox(height: 20,);
+          return SizedBox(
+            height: 20,
+          );
         },
       ),
       floatingActionButton: FloatingActionButton(
         onPressed: () {
           Navigator.push(
-            context,
-            MaterialPageRoute(
-              builder: (BuildContext context) =>
-                GenerateNews())
-          );
+              context,
+              MaterialPageRoute(
+                  builder: (BuildContext context) => GenerateNews()));
         },
         child: Icon(Icons.edit),
         backgroundColor: Color(0xff03AA5A),
