+++ lib/place/register_info.dart
